Add a flash module
